Refactor data-generator.js: move all the logic into index.js
